Move layout page code-behind codes into each ViewModel
