+++ src/Files.App/GlobalUsings.cs
@@ -49,7 +49,9 @@ global using global::Files.App.Data.Models;
 global using global::Files.App.Data.Parameters;
 global using global::Files.App.Data.TemplateSelectors;
 global using global::Files.App.UserControls;
+global using global::Files.App.UserControls.Selection;
 global using global::Files.App.ViewModels;
+global using global::Files.App.ViewModels.LayoutModes;
 global using global::Files.App.ViewModels.UserControls;
 global using global::Files.App.Views;
 global using global::Files.App.Views.LayoutModes;
