+++ src/Files.App/Views/LayoutModes/ColumnViewBase.xaml
@@ -1,5 +1,5 @@
 ﻿<!--  Copyright (c) 2023 Files Community. Licensed under the MIT License. See the LICENSE.  -->
-<local:StandardViewBase
+<local:BaseLayout
 	x:Class="Files.App.Views.LayoutModes.ColumnViewBase"
 	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
@@ -14,9 +14,10 @@
 	xmlns:wct="using:CommunityToolkit.WinUI.UI"
 	xmlns:wctconverters="using:CommunityToolkit.WinUI.UI.Converters"
 	x:Name="PageRoot"
+	DataContext="{x:Bind ViewModel}"
 	mc:Ignorable="d">
 
-	<local:StandardViewBase.Resources>
+	<local:BaseLayout.Resources>
 
 		<converters:StringToBrushConverter x:Key="StringToBrushConverter" />
 		<wctconverters:BoolNegationConverter x:Key="BoolNegationConverter" />
@@ -122,16 +123,16 @@
 			</Setter>
 		</Style>
 
-	</local:StandardViewBase.Resources>
+	</local:BaseLayout.Resources>
 
-	<Grid x:Name="RootGrid" ContextFlyout="{x:Bind BaseContextMenuFlyout}">
+	<Grid x:Name="RootGrid" ContextFlyout="{x:Bind ViewModel.BaseContextMenuFlyout}">
 
 		<!--  Empty Indicator  -->
 		<uc:FolderEmptyIndicator
 			HorizontalAlignment="Center"
 			VerticalAlignment="Center"
 			Canvas.ZIndex="0"
-			EmptyTextType="{x:Bind ParentShellPageInstance.FilesystemViewModel.EmptyTextType, Mode=OneWay}" />
+			EmptyTextType="{x:Bind ViewModel.ParentShellPageInstance.FilesystemViewModel.EmptyTextType, Mode=OneWay}" />
 
 		<!--  Invalid Item Name Tip  -->
 		<TeachingTip
@@ -145,17 +146,17 @@
 		<!--  Grouping/Ungrouping  -->
 		<SemanticZoom
 			x:Name="RootGridZoom"
-			CanChangeViews="{x:Bind CollectionViewSource.IsSourceGrouped, Mode=OneWay}"
+			CanChangeViews="{x:Bind ViewModel.CollectionViewSource.IsSourceGrouped, Mode=OneWay}"
 			ViewChangeStarted="SemanticZoom_ViewChangeStarted"
-			Visibility="{x:Bind FolderSettings.IsLayoutModeChanging, Mode=OneWay, Converter={StaticResource NegatedBoolToVisibilityConverter}}">
+			Visibility="{x:Bind ViewModel.FolderSettings.IsLayoutModeChanging, Mode=OneWay, Converter={StaticResource NegatedBoolToVisibilityConverter}}">
 
 			<!--  Main View  -->
 			<SemanticZoom.ZoomedInView>
 				<ListView
 					x:Name="FileList"
-					wct:ScrollViewerExtensions.EnableMiddleClickScrolling="{x:Bind IsMiddleClickToScrollEnabled, Mode=OneWay}"
+					wct:ScrollViewerExtensions.EnableMiddleClickScrolling="{x:Bind ViewModel.IsMiddleClickToScrollEnabled, Mode=OneWay}"
 					x:FieldModifier="public"
-					AllowDrop="{x:Bind InstanceViewModel.IsPageTypeSearchResults, Converter={StaticResource BoolNegationConverter}, Mode=OneWay}"
+					AllowDrop="{x:Bind ViewModel.InstanceViewModel.IsPageTypeSearchResults, Converter={StaticResource BoolNegationConverter}, Mode=OneWay}"
 					CanDragItems="True"
 					ContainerContentChanging="FileList_ContainerContentChanging"
 					DoubleTapped="FileList_DoubleTapped"
@@ -168,7 +169,7 @@
 					IsDoubleTapEnabled="True"
 					IsRightTapEnabled="True"
 					IsTabStop="True"
-					ItemsSource="{x:Bind CollectionViewSource.View, Mode=OneWay}"
+					ItemsSource="{x:Bind ViewModel.CollectionViewSource.View, Mode=OneWay}"
 					PreviewKeyDown="FileList_PreviewKeyDown"
 					RightTapped="FileList_RightTapped"
 					ScrollViewer.IsScrollInertiaEnabled="True"
@@ -405,7 +406,7 @@
 			<SemanticZoom.ZoomedOutView>
 				<ListView
 					HorizontalAlignment="Stretch"
-					ItemsSource="{x:Bind CollectionViewSource.View.CollectionGroups, Mode=OneWay}"
+					ItemsSource="{x:Bind ViewModel.CollectionViewSource.View.CollectionGroups, Mode=OneWay}"
 					SelectionMode="None">
 					<ListView.ItemTemplate>
 						<DataTemplate x:DataType="ICollectionViewGroup">
@@ -488,4 +489,4 @@
 		</Canvas>
 
 	</Grid>
-</local:StandardViewBase>
+</local:BaseLayout>
